requirements.txt
requests
beautifulsoup4
webdriver-manager
undetected-chromedriver
chromedriver-py
pyarrow
selenium
pandas
gspread
google-auth